% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{scrape_patient_forum_page}
\alias{scrape_patient_forum_page}
\title{Scrape page}
\usage{
scrape_patient_forum_page(url)
}
\arguments{
\item{url}{URL to the page to scrape}
}
\value{
A data frame
}
\description{
Convert patient info page into data frame
}
