% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{scrape_patient_forum_pages}
\alias{scrape_patient_forum_pages}
\title{Scrape pages}
\usage{
scrape_patient_forum_pages(urls)
}
\arguments{
\item{urls}{URLs to the page to scrape}
}
\value{
A list of data frames
}
\description{
Convert patient info pages into list of data frames
}
